Update and rename numberOf_subsets_with_given_diff.cpp to subsets_with_given_diff.cpp
